# api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Vitality
  version: v1
#host: my.remote.host
host: localhost:3000
basePath: /
schemes:
  - http
  - https
consumes:
- application/json
- application/javascript
produces:
- application/json
- application/javascript
- text/javascript
paths:
...
  /storecommitactivity:
    x-swagger-router-controller: core_storecommitactivity
    get:
      summary: Send commit activity to a persistent data store
      description: |
        Send commit activity to a persistent data store.
      operationId: storecommitactivity
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /getrawrepoactivity:
    x-swagger-router-controller: stats_getrawrepoactivity
    get:
      summary: Get all activity data for one or more repositories
      description: |
        Get all activity data for one or more repositories.  Ommitting the 'repo' parameter will return data for all available repositories.
      operationId: getrawrepoactivity
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
        - name: repo
          in: query
          description: Comma-seperated string of repository names
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getrawrepoactivityResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getrawcommitactivity:
    x-swagger-router-controller: stats_getrawcommitactivity
    get:
      summary: Get all commit data for one or more repositories
      description: |
        Get all commit data for one or more repositories.  Ommitting the 'repo' parameter will return data for all available repositories.
      operationId: getrawcommitactivity
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
        - name: repo
          in: query
          description: Comma-seperated string of repository names
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getrawcommitactivityResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /getcommitsbycommitter:
    x-swagger-router-controller: stats_getcommitsbycommitter
    get:
      summary: Get internal/external commits
      description: |
        Get internal/external commits.
      operationId: getcommitsbycommitter
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getcommitsbycommitterResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vis/getcommitsbyrepo:
    x-swagger-router-controller: visjs_getcommitsbyrepo
    get:
      summary: Get commit count for each repository by date, formatted for consumption by visJS
      description: |
        Get commit count for each repository by date.  Results are returned in a format that is readily consumable by VisJS.
      operationId: getcommitsbyrepo
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getcommitsbyrepoResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vis/getcommittertotals:
    x-swagger-router-controller: visjs_getcommittertotals
    get:
      summary: Get internal/external committer counts grouped by week, formatted for consumption by visJS
      description: |
        Get internal/external committer counts grouped by week.  Results are returned in a format that is readily consumable by VisJS.
      operationId: getcommittertotals
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getcommittertotalsResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vis/getcommittotals:
    x-swagger-router-controller: visjs_getcommittotals
    get:
      summary: Get internal/external commit counts grouped by week, formatted for consumption by visJS
      description: |
        Get internal/external commit counts grouped by week.  Results are returned in a format that is readily consumable by VisJS.
      operationId: getcommittotals
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getcommittotalsResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ErrorResponse:
    type: string
    required:
      - success
      - message
    properties:
      success:
        type: string
      message:
        type: string
      nativemessage:
        type: string
      nativecode:
        type: integer
  SuccessResponse:
    type: string
    required:
      - success
    properties:
      success:
        type: string
...
  ShortRepo:
    type: object
    required:
      - created
      - repoid
      - reponame
      - open_issues
      - forks
      - stargazers
      - watchers
    properties:
      created:
        type: string
      repoid:
        type: integer
      reponame:
        type: string
      open_issues:
        type: integer
      forks:
        type: integer
      stargazers:
        type: integer
      watchers:
        type: integer
...
  getrawrepoactivityResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: boolean
      data:
        type: array
        items:
          $ref: "#/definitions/ShortRepo"
  CommitActivity:
    type: object
    required:
      - reponame
      - date
      - name
    properties:
      reponame:
        type: string
      date:
        type: string
      name:
        type: string
  getrawcommitactivityResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: boolean
      data:
        type: array
        items:
          $ref: "#/definitions/CommitActivity"
...
  CommitterCountData2:
    type: object
    required:
      - name
      - ibm
      - reponame
      - date
    properties:
      name:
        type: string
      ibm:
        type: integer
      reponame:
        type: string
      date:
        type: string
  getcommitsbycommitterResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/CommitterCountData2"
...
  VisData1:
    type: object
    required:
      - y
      - x
      - group
    properties:
      y:
        type: integer
      x:
        type: string
      group:
        type: string
  getcommitsbyrepoResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/VisData1"
  VisData2:
    type: object
    required:
      - y
      - x
      - group
    properties:
      y:
        type: integer
      x:
        type: string
      group:
        type: integer
  getcommittertotalsResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/VisData2"
  getcommittotalsResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/VisData2"